refactor: move agent initialization into a hook, tidy migration check
