+++ packages/legacy/core/App/index.ts
@@ -53,6 +53,7 @@ import { CredentialListFooterProps } from './types/credential-list-footer'
 import InactivityWrapper, { AutoLockTime } from './components/misc/InactivityWrapper'
 import { OpenIDCredentialRecordProvider } from './modules/openid/context/OpenIDCredentialRecordProvider'
 import { defaultConfig } from './container-impl'
+import useInitializeAgent from './hooks/initialize-agent'
 
 export * from './navigators'
 export * from './services/storage'
@@ -77,7 +78,7 @@ export { createStyles } from './screens/OnboardingPages'
 export { statusBarStyleForColor, StatusBarStyles } from './utils/luminance'
 export { BifoldError } from './types/error'
 export { EventTypes } from './constants'
-export { didMigrateToAskar, migrateToAskar } from './utils/migration'
+export { migrateToAskar } from './utils/migration'
 export { createLinkSecretIfRequired, getAgentModules } from './utils/agent'
 export { removeExistingInvitationIfRequired, connectFromScanOrDeepLink } from './utils/helpers'
 
@@ -161,6 +162,7 @@ export {
   OpenIDCredentialRecordProvider,
   NotificationListItem,
   useDefaultStackOptions,
+  useInitializeAgent,
   Splash,
   Developer,
   Terms,
